Crate font8x8

8x8 monochrome bitmap fonts for rendering

This crate contains a Rust implementation of a public domain 8-bit by 8-bit font.

See ’../README.md`.

On the miscellanous characters included

These characters are provided as-is, and some don’t have a proper unicode point. They are provided in the MISC and SGA contants. For a description, please read the constant documentation.

Credits

Initially extracted from an asm file, fetched from a now broken link: http://dimensionalrift.homelinux.net/combuster/mos3/?p=viewsource&file=/modules/gfx/font8_8.asm

It was then ported to a C header file, also in the Public Domain, https://github.com/dhepper/font8x8.

This crate is an extension of that work.

Re-exports

pub use self::unicode::FontUnicode;

pub use self::unicode::UnicodeFonts;

Modules

legacy

Re-export the original [u8; 8] constants, taken from C-header files. Legacy constants.

unicode

Unicode support for 8x8 fonts.

Constants

BASIC_FONTS

A convenient constant for Basic Latin fonts (U+0000 - U+007F), that implements the UnicodeFonts trait.

BASIC_UNICODE

A constant [FontUnicode; 128], for Basic Latin fonts (U+0000 - U+007F).

BLOCK_FONTS

A convenient constant for Block Element fonts (U+2580 - U+259F), that implements the UnicodeFonts trait.

BLOCK_UNICODE

A constant [FontUnicode; 32], for Block Element fonts (U+2580 - U+259F).

BOX_FONTS

A convenient constant for Box Element fonts (U+2500 - U+257F), that implements the UnicodeFonts trait.

BOX_UNICODE

A constant [FontUnicode; 128], for Box Element fonts (U+2500 - U+257F).

GREEK_FONTS

A convenient constant for Greek fonts (U+0390 - U+03C9), that implements the UnicodeFonts trait.

GREEK_UNICODE

A constant [FontUnicode; 128], for Greek fonts (U+0390 - U+03C9).

HIRAGANA_FONTS

A convenient constant for Hiragana fonts (U+3040 - U+309F), that implements the UnicodeFonts trait.

HIRAGANA_UNICODE

A constant [FontUnicode; 96], for Hiragana fonts (U+3040 - U+309F).

LATIN_FONTS

A convenient constant for Extended Latin fonts (U+00A0 - U+00FF), that implements the UnicodeFonts trait.

LATIN_UNICODE

A constant [FontUnicode; 96], for Extended Latin fonts (U+00A0 - U+00FF).

MISC_FONTS

A convenient constant for Miscellanous fonts (U+20A7, U+0192, U+00AA, U+00BA, U+2310, U+2264, U+2265, U+0060, U+1EF2, and U+1EF3), that implements the UnicodeFonts trait.

MISC_UNICODE

A constant [FontUnicode; 10], for Miscellanous fonts (U+20A7, U+0192, U+00AA, U+00BA, U+2310, U+2264, U+2265, U+0060, U+1EF2, and U+1EF3).

SGA_FONTS

A convenient constant for special SGA fonts (U+E541 - U+E55A), that implements the UnicodeFonts trait.

SGA_UNICODE

A constant [FontUnicode; 26], for special SGA fonts (U+E543 - U+E55A).